// src/context/type.ts
export enum Type {
  SET_ACTIVE = 'SET_ACTIVE',
  SET_TASK = 'SET_TASK',
  ADD_TASK = 'ADD_TASK',
  SET_TASK_FILTER = 'SET_TASK_FILTER',
  SET_DONE = 'SET_DONE',
  SET_TASK_COMPLETED = 'SET_TASK_COMPLETED',
  SET_NO_DONE = 'SET_NO_DONE',
  SET_USER = 'SET_USER',
  SET_TASKS = 'SET_TASKS',
  SET_EMPTY_TASKS = 'SET_EMPTY_TASKS',
  CLEAR_DATA = 'CLEAR_DATA',
  SET_LAN = 'SET_LAN',
  SET_ALERT = 'SET_ALERT',
  SET_BACKGROUND_ID = 'SET_BACKGROUND_ID',
  SET_SORT_TYPE = 'SET_SORT_TYPE',
  DELETE_TASK = 'DELETE_TASK',
}